celeb7.1

+++ controllers/cards.js
@@ -40,15 +40,11 @@ const createCard = (req, res) => {
 const deleteCard = (req, res) => {
   const { id } = req.params;
 
-  Card.findOneAndRemove({ _id: id })
+  Card.findOneAndRemove({ _id: id, owner: req.user._id })
     .then((card) => {
       console.log(card);
       if (!card) {
         return res.status(404).send({ message: 'Карточка с указанным _id не найдена' });
-      }
-      console.log(card.owner, req.user._id);
-      if (card.owner !== req.user._id) {
-        return res.status(403).send({ message: 'Недостаточно прав!' });
       }
       return res.send({ message: 'Карточка успешно удалена' });
     })
